python/requirements.txt
mongoengine==0.21.0
python-decouple==3.3
xmltodict==0.12.0
python-dateutil==2.8.1
Django==3.1.3
psycopg2==2.8.6
django-extensions==3.0.9
git+git://github.com/thejoeejoee/django-bulk-update-or-create.git@da833a4cea8a2026151c58a74b414e980353aade#egg=django-bulk-update-or-create